multi-console: move code into correct packages
